app/requirements.txt
Flask==3.0.3
Flask-Login==0.6.3
Flask-WTF==1.2.1
Flask-Bcrypt==1.0.1
Flask-Limiter==3.9.2
SQLAlchemy==2.0.32
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
